pyproject.toml
[project]
name = "qa-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.29.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
